requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions == 1.12.0
snowflake-snowpark-python[pandas] == 0.12.0
snowflake-connector-python[pandas]
requests == 2.28.1
pandas == 1.4.4